More tournament work. Split ActivityManager.
