+++ config/doctrine/ActivityParticipant.orm.xml
@@ -13,6 +13,7 @@
 		<many-to-one field="character" target-entity="Character" inversed-by="activity_participation"/>
 		<many-to-one field="style" target-entity="Style"/>
 		<many-to-one field="weapon" target-entity="EquipmentType"/>
+		<many-to-one field="armor" target-entity="EquipmentType"/>
 		<many-to-one field="group" target-entity="ActivityGroup" inversed-by="participants"/>
 		<one-to-one field="related_action" target-entity="Action" mapped-by="target_activity_participant"/>
 	</entity>
