+++ src/Command/ProcessActivitiesCommand.php
@@ -3,6 +3,7 @@
 namespace App\Command;
 
 use App\Service\ActivityManager;
+use App\Service\ActivityRunner;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
@@ -11,7 +12,7 @@ use Symfony\Component\Console\Output\OutputInterface;
 class ProcessActivitiesCommand extends Command {
 	private mixed $ruleset;
 
-	public function __construct(private ActivityManager $am) {
+	public function __construct(private ActivityRunner $ar) {
 		$this->ruleset = $_ENV['ACTIVITY_RULESET'];
 		parent::__construct();
 	}
@@ -24,7 +25,7 @@ class ProcessActivitiesCommand extends Command {
 	}
 
 	protected function execute(InputInterface $input, OutputInterface $output): int {
-		$this->am->runAll($this->ruleset);
+		$this->ar->runAll($this->ruleset);
 		$output->writeln("...activities complete");
 		return Command::SUCCESS;
 	}
